WIP
Implemented:
* HARD RESET
* MEMORY Unlock

Fixed:
* RS485: Fixed bug in RS485::send() that caused the function to return
  before the message was sent.
* watchdog timer: Fixed bug that caused the watchdog timer to reset
    during flash update

Optimized:
* tx sending speed
* power consumption

TODO:
* Check queue capacity before adding new element to queue in
    RS485::sendData()
* Add error handling to RS485::sendData()
* Add TX IRQ handler to uart

+++ include/Memory.h
@@ -62,7 +62,7 @@ static float* maxPv3        = (float *)(mainRegister + MAX_PV3_OFFSET);
 std::array<float*, 4> maximumValues = {maxPv0, maxPv1, maxPv2, maxPv3};
 
 /* ### READ ONLY VALUES ### */
-static uint8_t* memUnlocked   = (uint8_t *)(mainRegister + MEM_UNLOCKED_OFFSET);
+static uint32_t* memUnlocked   = (uint32_t *)(mainRegister + MEM_UNLOCKED_OFFSET);
 
 static uint64_t* error      = (uint64_t *)(mainRegister + ERROR_OFFSET);
 static uint64_t* status     = (uint64_t *)(mainRegister + STATUS_OFFSET);
